throw more Exception

+++ DFQXYU_HFT_2021221.Logic/CustomerLogic.cs
@@ -62,11 +62,19 @@ namespace DFQXYU_HFT_2021221.Logic
 
         public void Delete(int id)
         {
+            if (id < 1)
+            {
+                throw new IndexOutOfRangeException("Minimum id value is 1");
+            }
             this.customerRepo.Delete(id);
         }
 
         public Customer Read(int id)
         {
+            if (id < 1)
+            {
+                throw new IndexOutOfRangeException("Minimum id value is 1");
+            }
             return this.customerRepo.Read(id);
         }
