+++ DFQXYU_HFT_2021221.Logic/MovieRentalLogic.cs
@@ -110,11 +110,19 @@ namespace DFQXYU_HFT_2021221.Logic
 
         public void Delete(int id)
         {
+            if (id < 1)
+            {
+                throw new IndexOutOfRangeException("Minimum id value is 1");
+            }
             this.movieRentalRepo.Delete(id);
         }
 
         public MovieRental Read(int id)
         {
+            if (id < 1)
+            {
+                throw new IndexOutOfRangeException("Minimum id value is 1");
+            }
             return this.movieRentalRepo.Read(id);
         }
